feat: ✨ user dropdown in mobile header
